Resolving duplicated code on model employee and model health team
